+++ client/src/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import {apiFail} from './components/ListPlayers';
+import {apiFail} from './hooks/useAPI';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
